% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{ProvisioningClient-class}
\alias{ProvisioningClient}
\alias{ProvisioningClient-class}
\title{Client for working with Keboola Connection Provisioning API.}
\description{
The client allows to create and retrieve credentials for sandbox and
transformation buckets. Additional methods to kill sandbox/transformation
bucket are available.
}

