% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{ProvisioningClient-class}
\alias{ProvisioningClient}
\alias{ProvisioningClient-class}
\title{Client for working with Keboola Connection Provisioning API.}
\description{
The client allows to create and retrieve credentials for sandbox and
transformation buckets. Additional methods to kill sandbox/transformation 
bucket are available.
}
\section{Methods}{

\describe{
\item{\code{decodeResponse(response)}}{Internal method to process API response.
\subsection{Parameters}{\itemize{
\item{\code{response} List as returned from \code{httr} POST/GET methods.}
}}
\subsection{Return Value}{Response body - either list or string in case the body cannot be parsed as JSON.}}

\item{\code{dropCredentials(id)}}{Delete credentials and cleanup the associated database bucket.
\subsection{Parameters}{\itemize{
\item{\code{id} Credentials id.}
}}
\subsection{Return Value}{TRUE}}

\item{\code{getCredentials(type = "transformations")}}{Get new credentials of the given type or reuse existing credentials if available.
\subsection{Parameters}{\itemize{
\item{\code{type} Credentials type - either \code{transformations} or \code{sandbox}}
}}
\subsection{Return Value}{List with credentials.}}

\item{\code{getCredentialsById(id)}}{Get existing credentials.
\subsection{Parameters}{\itemize{
\item{\code{id} Credentials id.}
}}
\subsection{Return Value}{List with credentials.}}

\item{\code{initialize(backend, token, runId = "",
  url = "https://syrup.keboola.com/provisioning")}}{Constructor.
\subsection{Parameters}{\itemize{
\item{\code{backend} Type of database backend - either \code{mysql} or \code{redshift}}
\item{\code{token} KBC Storage API token.}
\item{\code{runId} Optional Run ID of the parent job.}
\item{\code{url} Optional URL of the provisioning API.}
}}}

\item{\code{killProcesses(id)}}{Kill database processes running with the given credentials.
\subsection{Parameters}{\itemize{
\item{\code{id} Credentials id.}
}}
\subsection{Return Value}{TRUE}}
}}

